# openapi.yaml
openapi: 3.1.0
info:
  title: Groceries API
  version: 1.0.0
  description: |
    REST API for the Groceries application. Provides token-based authentication
    suitable for native app clients and future frontend rewrites.

    ## Authentication

    All endpoints except `POST /api/v1/auth/login` require a Bearer token in the
    `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained from the login endpoint and expire after 720 hours (30 days).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          examples:
            - "item not found"

    # --- Auth ----------------------------------------------------------------
...
    ListItemSummary:
      type: object
      description: Presence of this object on an Item indicates the item is on the shopping list
      required: [id, quantity, done]
      properties:
        id:
          type: integer
          description: ID of the list entry (not the item itself)
          examples:
            - 42
        quantity:
          type: string
          examples:
            - "2 lbs"
        done:
          type: boolean
          description: Whether this item has been picked up during the current shopping trip
    Item:
      type: object
      required: [id, category_id, category_name, name]
      properties:
        id:
          type: integer
          examples:
            - 1
        category_id:
          type: integer
          examples:
            - 3
        category_name:
          type: string
          examples:
            - "Produce"
        name:
          type: string
          examples:
            - "Apples"
        list:
          $ref: "#/components/schemas/ListItemSummary"
          description: Present when this item is currently on the shopping list
    CreateItemRequest:
      type: object
      required: [category_id, name]
      properties:
        category_id:
          type: integer
          examples:
            - 3
        name:
          type: string
          minLength: 1
          examples:
            - "Apples"
    UpdateItemRequest:
      type: object
      required: [category_id, name]
      properties:
        category_id:
          type: integer
          examples:
            - 3
        name:
          type: string
          minLength: 1
          examples:
            - "Apples"

    # --- Shopping list -------------------------------------------------------
...
    RecipeItem:
      type: object
      required: [id, recipe_id, item_id, item_name, quantity, in_list]
      properties:
        id:
          type: integer
          examples:
            - 7
        recipe_id:
          type: integer
          examples:
            - 2
        item_id:
          type: integer
          examples:
            - 1
        item_name:
          type: string
          examples:
            - "Apples"
        quantity:
          type: string
          examples:
            - "3"
        in_list:
          type: boolean
          description: Whether this item is currently on the shopping list
    Recipe:
      type: object
      required: [id, name, description, created_at, items]
      properties:
        id:
          type: integer
          examples:
            - 2
        name:
          type: string
          examples:
            - "Apple Pie"
        description:
          type: string
          examples:
            - "Classic American apple pie"
        created_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecipeItem"
...
    UpdateRecipeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          examples:
            - "Apple Pie"
        description:
          type: string
    AddRecipeItemRequest:
      type: object
      required: [item_id]
      properties:
        item_id:
          type: integer
          examples:
            - 1
        quantity:
          type: string
          default: ""
          examples:
            - "3"
...
  responses:
    NoContent:
      description: Operation succeeded with no response body
    BadRequest:
      description: The request was malformed or failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The operation conflicts with existing data (e.g. duplicate name)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: An unexpected server-side error occurred
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Numeric resource ID
    ItemIdPath:
      name: itemId
      in: path
      required: true
      schema:
        type: integer
      description: Numeric item ID

# ---------------------------------------------------------------------------
# Default security (overridden on the login endpoint)
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/items:
    get:
      operationId: listItems
      summary: List all items
      tags: [items]
      parameters:
        - name: category_id
          in: query
          schema:
            type: integer
          description: Filter items by category
        - name: in_list
          in: query
          schema:
            type: boolean
          description: Filter to only items that are (or are not) on the shopping list
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      operationId: createItem
      summary: Create a new item
      tags: [items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/items/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"

    get:
      operationId: getItem
      summary: Get an item by ID
      tags: [items]
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateItem
      summary: Update an item
      tags: [items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteItem
      summary: Delete an item
      description: Fails if the item is used in any recipes.
      tags: [items]
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --------------------------------------------------------------------------
  # Shopping list
  # --------------------------------------------------------------------------
...
  /api/v1/list/finish:
    post:
      operationId: finishShopping
      summary: Finish shopping - remove all done items from the list
      tags: [list]
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --------------------------------------------------------------------------
  # Recipes
  # --------------------------------------------------------------------------
...
  /api/v1/recipes/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"

    get:
      operationId: getRecipe
      summary: Get a recipe by ID, including its items
      tags: [recipes]
      responses:
        "200":
          description: Recipe with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateRecipe
      summary: Update a recipe's name and description
      tags: [recipes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipeRequest"
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteRecipe
      summary: Delete a recipe
      tags: [recipes]
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/recipes/{id}/items:
    parameters:
      - $ref: "#/components/parameters/IdPath"

    post:
      operationId: addRecipeItem
      summary: Add an item to a recipe
      tags: [recipes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRecipeItemRequest"
      responses:
        "201":
          description: Item added to recipe; returns the full updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Item is already in the recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/recipes/{id}/items/{itemId}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
      - $ref: "#/components/parameters/ItemIdPath"

    delete:
      operationId: removeRecipeItem
      summary: Remove an item from a recipe
      tags: [recipes]
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...